requirements.txt
pytorch=2.1.1
gensim==4.3.2
tqdm==4.66.1
wandb==0.16.0
numpy==1.26.2
pandas==2.1.3
matplotlib==3.8.2
scikit-learn==
scipy==1.3.2
spacy==3.7.2
ipykernel==6.26.0
tensorboard==2.15.1